templatize `savanna_cluster` by `NUM_NODES`
